move initializer_list to its own file
This makes it easier to update the copy of this code embedded in the
ArxContainer library, since now the entire file can be copied as-is.

This commit changes the preprocessor guard that was around the the
definition of initializer_list, to protect it from being double-defined
by ArxContanier as well. It is now a standard include guard named after
the file. This might cause problems when combined with ArxContainer,
which should be solved as soon as ArxContainer also includes this same
file.
