+++ ArxTypeTraits/type_traits.h
@@ -367,28 +367,9 @@ namespace std {
         b = move(t);
     }
 
-
-#ifndef ARX_TYPE_TRAITS_INITIALIZER_LIST_DEFINED
-#define ARX_TYPE_TRAITS_INITIALIZER_LIST_DEFINED
-
-    template<class T>
-    class initializer_list
-    {
-    private:
-        const T* array;
-        size_t len;
-        initializer_list(const T* a, size_t l) : array(a), len(l) {}
-    public:
-        initializer_list() : array(nullptr), len(0) {}
-        size_t size() const { return len; }
-        const T *begin() const { return array; }
-        const T *end() const { return array + len; }
-    };
-
-#endif // ARX_TYPE_TRAITS_INITIALIZER_LIST_DEFINED
-
 } // namespace std
 
+#include "initializer_list.h"
 #include "tuple.h"
 #include "functional.h"
